Define route pattern
This pattern is adapted from the Sustainability Toolkit project[1] and
mirrors several best practices in defining Express routes and route
handlers.

This also introduces a new tool for our tests: supertest[2].  Supertest
is the canonical way to test express services and makes it possible to
test a service directly / without having to actually run it.

[1] https://gitlab.com/dial/principles/sustainability-toolkit-api/-/tree/development
[2] https://www.npmjs.com/package/supertest

Issue #16 Add a couple of routes

Co-authored-by: Jesse Bickel <bickelj@gmail.com>