Add Jest + Tests
Jest is a testing framework for NodeJS.  It requires some extra packages
to be able to run in TypeScript, so we're including those here too.

This also runs tests on CI and integrates with CodeCov -- CodeCov is
proprietary, but free.It adds some nice coverage analysis as part of
our CI workflow.  I figure it is OK to use here because it is not in any
required to contribute to / operate this code base.

Issue #4