[DRAFT] DB support
This commit does a lot -- it adds support for database interaction in
addition to adding migrations for the schema defined in #2.

It may be we want to do this all at once; but it also may be that we
want to spread some of these migrations out as they are actually needed.
The benefit of that approach is that we can review the schema in the
context in which it will be used.

Issue #8