Add database support
We are using a few packages as part of our database integration:

1. `tinypg` is very light weight tool for connecting to and querying
   a postgres database.
2. `postgres-migrations` is a similarly light migration tool.  It has
   a philosophy of "no rollbacks" which makes life quite easy.

I'm also adding a generally useful library `dotenv` which makes it easy
to pull environment variables from a `.env` file if they aren't already
set by the system.

In addition to setting up the packages this commit also creates the db
objects that will be used.  Migrations will be stored in the
`src/database/migrations` directory and the various query objects will
be stored in `src/database/queries`

Issue #8 Set up schema