Add action to build and push docker images to GHCR
This helps enable Continuous Integration and Continuous Delivery, maybe
even Continuous Deployment, by automatically pushing deployable images.

A development/latest docker image will be available at:

    ghcr.io/philanthropydatacommons/server:latest

Alternatively, specific versions of images will be available at:

    ghcr.io/philanthropydatacommons/server:[version]

The version scheme is based entirely on the commit from which the docker
image was built. It is the date of the commit followed by the short hash
of the commit, not the date the build was run. In this way, the version
is deterministic (assuming a clean repository workspace), friendly-ish
to humans (the date portion), unique and unlikely to collide (the short
hash within a single day is not as likely to collide versus using either
the date or only the short hash of the commit), and relatively short.

For example, the version 20220630-abcd123 would indicate that the image
was built from a commit made on June 30, 2022 and having a git commit id
starting with abcd123.

Issue #3 Set up CI / CD