Notify a workflow in deploy project on image push
Use workflow dispatch to call a manual-run workflow in the 'deploy'
project to tell it the image version (docker image tag) that was just
pushed to GHCR. The workflow in 'deploy' in turn updates the deployment
scripts. At the moment, this unfortunately uses a Personal Access Token
(PAT), see the README.md in the .github/workflows directory.

Issue #3 Set up CI / CD