Remove ts-node dependency for migrations
We have `ts-node` as a dev dependency, but migrations will be run in
a production environment that should NOT rely on dev dependencies.

This adjusts our migrations to run using the transpiled version of our
code base. In making this change I discovered an issue with tsm where it
has no ability to copy files that aren't typescript files.  This is
a questionable decision on the part of the tsm developers but we have to
work around it.

copyfiles appears to be the right tool for cross-platform file copying,
which is just to say there is a new build step that copies over our
.keep and our .sql files into the `dist` directory.