Use 'buildcache' for docker image build cache
Instead of re-purposing 'latest' tag, use 'buildcache' tag.
Since we deploy with linux/amd64, make that explicit as well.

The cache-from and cache-to which we now use with label/tag buildcache
as I understand it makes a docker build faster by avoiding pulling the
base image every time. In other words, I think GitHub's Container
Registry will have our previous image layers locally, and whenever the
first few layers are identical (majority case), it will be much faster
to re-use them.

The reason to use a separate tag for buildcache is that apparently it is
not a pullable/usable artifact with docker pull. The reason to keep the
latest tag is convenience for a one-off or exploratory pull.

Issue #39 no matching manifest for linux/amd64 on docker pull ...:latest