Add pino logging libraries
Because pino was easy to get started with and claims speed, try it out.

Pino has a request logging library called pino-http. The logs are in
JSON format by default and unix times since epoch as well, but there is
also the pino-pretty library that will make these friendly in terminals.

We can pipe to pino-pretty on our ":dev" npm tasks such as start:dev and
migrate:dev. When doing so, use RFC3339 datetime format when prettyfying
because it is a good, useful standard. Production logs can be sent to it
later, I assume.

Issue #41 Add a logger