Add GET /canonicalFields endpoint
This adds support for retrieving all canonical fields contained in the
system.  As part of this change we:

1. Introduce `ajv` which is a runtime validator that uses JSONSchema in
   order to ensure that data conforms to a given TypeScript type.
2. Introduce migrations and basic queries for the `canonical_field`
   table.
3. Introduces a new pattern for integration tests that rely on fixture
   data.

We want to be sure the data result conforms to our API spec / expected
type output so we actually validate it before passing it along to the
response middleware.

Note that Express route handlers cannot be asynchronous which is why we
have to use the promise `then` and `catch` syntax.

Issue #30 Implement GET /canonicalFields API endpoint