Migrate the database schema on container startup
When doing two or more commands on startup, it is straightforward to use
a small script. The downside is the command is not visible directly in
the Dockerfile, there is one level of indirection. But we do want to
do `npm run migrate` then `npm run start`. The migrate task should be
idempotent and therefore harmless to run on repeated startup even with
the same version of the software (e.g. no new schema changes).

Issue #56 Do DB migration on startup