Require named loggers
We want to make it easy for files to create loggers that point back to
the file generating a given log.

By exposing ONLY `getLogger` method we support a cleaner for the logger
syntax for the logger (rather than importing `logger` and having to use
`childLogger` in our code, we can just import `getChildLogger` and
assign the child logger to a local `logger` value).

This also ensures that all loggers are generated with a name (I suppose
someone could choose to name with an empty string).