Use env var `LOG_LEVEL` to set pino log level
The environment variable `LOG_LEVEL` will set the level on the root
logger when it is set. When running tests under jest, the globalSetup
examines the (jest) global and project configurations as well as the env
var `LOG_LEVEL` to decide whether to silence pino using the same env var
`LOG_LEVEL`. If you set `LOG_LEVEL`, that level takes precedence in all
environments. If you do not set `LOG_LEVEL` and the jest config sets
`silent: true`, the globalSetup for jest sets `LOG_LEVEL` to `silent` on
your behalf, essentially passing jest's `silent: true` setting through.

There are multiple ways to set `LOG_LEVEL` such as on the command line
or in a `.env` file.

Issue #59 Get pino and jest to agree on console verbosity