Draft 2 of GET /opportunities with test and types
This is a progress commit on the /opportunities service endpoint. The
test inserts dummy data sufficient to test getting a full opportunity.
The application schema will be inferred for now from whatever is
posted to the /opportunities, that is why the application schema and
fields are here. It could be inferred from another endpoint, though,
such as /applications, if we want, or of course made explicit with its
own endpoint such as /applicationSchemas.

Issue #31 Add GET /opportunities