Draft/progress commit
The query to fetch all is wide and returns quite a bit of repetition.
The trick will be to map/reduce/filter/foreach/whatever the rows down
into the desired JSON schema. Alternatively, we might find a postgres
JSON function that could be used to do much of this work on the db side.

Issue #31 Add GET /opportunities