Enable coercion in AJV
By default AJV will not attempt to coerce values (e.g. if a value is
"42" and it is expected to be an integer, ajv will flag that as
a string).  This is a good default but makes it hard to validate things
like HTTP bodies and query parameters, since Express passes those as
strings regardless.

Enabling coercion means that we can validate numbers passed in the
querystring.  Coercion will also modify the original data so that it
results in the correct type[1]!  This is critical since without that
modification we would lose type safety.

[1] https://ajv.js.org/guide/modifying-data.html