Support unique error detection
We need to know when an insert query fails due to a unique constraint
violation so that we can share the appropriate error to the end user
without exposing too much detail about our backend.

The `tinypg` library does not have strong types when errors are thrown,
so in order to do that detection safely we need to essentially patch the
library with a type of our own.