Define schema for query parameters
We tried using a slightly consolidated way of defining an ajv validator
for the query parameters of the GET opportunity endpoint.  It turns out
that consolidation came at the cost of type validation which is frankly
worth the extra line of verbosity.

By defining the schema directly with a type we will be warned by the
static analysis if our schema doesn't match the expected type.