Add POST /applicants endpoint
Applicants need to be registered in PDC before related data can be
stored against them. This endpoint enables that behavior.

Right now the endpoint does not allow an applicant to be created with
the opted in status. This is an attempt to discourage developers from
setting a "true" flag without actually getting the user's permission.