Add POST /applicationForms endpoint
This endpoint allows a client to register a new application form for an
opportunity.  This implementation of the endpoint is complete from
a pure REST entity sense but it may be burdensome for an end user to
have to register each field within the form as individual entities / API
calls.

It is likely that this endpoint will be expanded to support the creation
of fields alongside the creation of the form.

Issue #64