Add ApplicationFormFields
When a user creates a new ApplicationForm we want them to also define
the ApplicationFormFields associated with the new form.  This means the
`POST /applicationForms` endpoint needs to take a `fields` attribute.

Long ago we had punted on the complexity of multi-entity API calls, but
this really required shaving that yak and so this commit does so.  For
example:

1. We now have the concept of a `Write` version of a given entity.  This
   is in order to fill a feature void.  In JsonSchema the idea of
  `readonly` lets us define a single schema that works for POST and GET
  because the readonly fields are ignored on POST.  TypeScript has no
  such luxury.
2. In order to truly support the "modified" versions of types for Write,
   we want to be able to use Generics to create type aliases.  This
   unfortunately requires a linter change, and we should be very
   critical of that change.
3. TypeScript doesn't provide a generic for "swapping" the type of
   a given interface attribute, as far as I could find.  This means
   I had to also enable literals in our linter in order to use the
   "Write" version of ApplicationFormField in the ApplicationFormWrite
   type.

Now that we are defining `Write` types we can remove certain handler
interfaces (ones that truly just take in a `Write` version of a type.