Refactor error handling
We had been implementing error handling for our endpoints on an
endpoint-by-endpoint basis, which means each endpoint had to select the
correct message / status code / etc for a given circumstance. This
approach risks inconsistencies and was starting to result in a fair bit
of duplicated logic across different handlers.

Express makes it possible to define error-handling middleware[1].
Essentially what happens is that handlers can pass an error to the
`next` step in the middleware stack and express will route the logic to
our custom handler.  The handler can then resolve the error to the
correct status code and response body.

As part of this new approach we will communicate the nature of a given
"error" by defining custom error types for different types of error.
For instance an internal validation error should be handled differently
from an input validation error, and so the two errors warrant having
separate classes.

[1] https://expressjs.com/en/guide/error-handling.html