Set up CI test configuration
We have unit tests and integration tests; our CI was set up to run each
independently and attempt to merge the resulting code coverage.  This is
a fine concept in theory but it results in two problems:

1. The coverage merging was strange and, seemingly, buggy.  This meant
   the CI coverage checks were behaving unpredictably.
2. There isn't any way to generate the same coverage report that CI uses
   when running locally!  This makes it hard to know how to update
   tests to ensure proper coverage.

Keeping the separate unit / integration test comands is still valuable
because there are situations where coverage isn't needed (e.g. when just
verifying that local tests run), and these commands are slightly faster
as they don't run migrations for the unit tests.

Issue #98