Create GET /proposals endpoint
This creates a new entity type of `Proposal` and defines the first
endpoint for accessing proposal data.

We had been talking about `Application` instead of `Proposal` but that
term could be confused to mean the "Application Form" rather than
a given set of data submitted in order to apply.

Issue #34 Implement GET /applications endpoint