Add POST /proposals endpoint
This endpoint creates a new proposal for a given applicant + opportunity
combination.  An applicant can have more than one proposal for a given
opportunity but distinct proposals must be uniquely identifiable by an
external ID (this allows future versioning.)

At this point creating a proposal is NOT the same as storing values
associated with a proposal. A separate POST to `/proposalVersions` (not
implemented yet) will need to be made in order to store proposal data.

Issue #35