Fix tests involving multiple DB connections
It turns out the way the schema was being set was only applying to
a single connection from the connection pool (in retrospect this makes
sense, looking at the code).  We need ALL connections invoked by a given
test to point to the correct schema.

By setting the correct environment variable future connections will be
kicked off properly.

This approach has the downside that test runners cannot set PGOPTIONS
when running tests.

If this ends up being a requirement for anybody we'll need to find
a way to amend PGOPTIONS instead of overriding it.