Update error return schema
We were returning an error with a `message` and an `errors` property.
This changes the `errors` property to a more flexible `details`
property.  We were using `errors` before because in the case of Database
errors the details happened to be a complex type of object that our
validation library called an "Error".

This also documents the "name" property which has existed for a while
now.