Add a script to build and post applicationForm
The examples are not updated here but the new script is called
buildApplicationForm.sh. It takes an xlsx file and url and performs
all the steps to make an application form and post it to the PDC's
/applicationForm endpoint. It appears to work for the two spreadsheets
in the issues #94 and #95.

Issue #102 Make canonical fields and application forms from sheets
Issue #94 Opportunity data fields for the MacArthur Foundation/Fluxx
Issue #95 Opportunity data fields for the Miami Foundation/Submittable