Require safer rules for arrays
TypeScript, by default, treats all array index access as not being
undefined.  This in itself is a little unsafe -- there's no guarantee
that foo[42] is NOT undefined -- but what makes it really rough is that
it becomes a lint error if you *want* to check that the value is
defined.

```
const foo = ['hello'];
const bar = foo[42];
if (bar === undefined) // the linter will complain here.
 ...
```

The solution to this problem is to tell TypeScript that array access CAN
result in an undefined value / require index access to be checked[1].

We weren't relying on unchecked indexes outside of our tests, but this
change requires us to be safer there as well (which is a good thing,
really).

[1] https://www.typescriptlang.org/tsconfig#noUncheckedIndexedAccess