Fix OpenAPI spec for ApplicationFormField
A few iterations back I had explored the idea of using
a canonicalFieldShortCode as a foreign key instead of
a canonicalFieldId.  We ultimately landed on using the Id instead, but
apparently I missed one of the old values when updating the open API
spec.

This corrects that mistake to be in line with our intent.

Issue #110