For audit capabilities, log subset of API keys
When using x-api-key for auth, there is only a set of authentication
keys. To have a clearer picture of which entity made which API call,
log some subset of that key, assuming that it is long enough to not
compromise it by this logging.

Issue #135 Logging and creating an audit trail