Add option to deep GET /proposals/{proposalId}
When optional query parameter includeFieldsAndValues is set to true on
GET /proposals/{proposalId}, include all proposal versions, associated
values, and associated application form fields in the response.

By returning this almost-fully-deep object tree, it is more convenient
for the caller and more efficient in terms of request, query, and
response counts. The assumption is that the purpose of the PDC is to
see and compare which fields are used for what purpose for proposals.

This commit consolidates incremental changes that should be seen here:
https://github.com/PhilanthropyDataCommons/service/pull/210

Issue #101 Implement GET /proposals/{proposalId} endpoint