Use a db migration script to seed canonical fields
With this approach, many canonical fields will already be present in
the database following database migration. This makes it easier to get
started working on the PDC service software by having portable
canonical field ids, i.e. ids are consistent across the official
instance of PDC and also local development instances. It also lowers
the barrier to creation of an application form.

If we do not have canonical fields present, it is more challenging to
generate an application form, e.g. from a spreadsheet. This approach
assumes that an original spreadsheet can be viewed and/or edited by
folks with domain knowledge and then used to generate migration
scripts as well as application form creation JSON. In fact this is how
the initial SQL script here was created.

This approach replaces the previous attempt at fully automatic
generation of canonical fields from unedited spreadsheets of fields.
As such it is a kind of breaking change at the database data level. In
order for migration to succeed with this change, an empty
canonical_fields table is expected. So before running migrate, run a
'delete from canonical_fields;' query.

As time goes on and fields are modified, two approaches could make
sense:
1. Refine the initial insert script and have it delete all the data in
the database and re-insert it. This is the easier approach, assuming
all form fields and other data can be re-inserted via automation.
2. Use additional scripts to carefully update and preserve data. This
approach will need to be taken at some point in the future once more
data is present.

This commit takes approach (1) but could potentially take (2), e.g.
start with what is present in the pilot PDC instance and carefully
insert and update where appropriate (with additional work).

See the script used to generate the canonical fields SQL file here:
https://github.com/PhilanthropyDataCommons/data-scripts/blob/ba13a1d4b92319cc72f0d96e1c5fa888bb56088d/src/generateCanonicalFieldsInserts.ts#L1

Issue #102 Make canonical fields and application forms from spreadsheets