Configure CORS
We expect clients, including the data-viewer, to make requests from a
different domain than the API itself. For security, browsers will not
allow cross-origin requests[1] that include credentials - as are required
by every endpoint of this API - in order to protect against malicious
third-party websites from exploiting a logged-in user's session.

Install and use the CORS middleware[2] with the default options, which
allows requests from any origin. This is not the ideal security policy
from a defense-in-depth standpoint, but is useful during development.
The current lack of destructive endpoints is one mitigation, and the
lack of cookie authentication is another: browser requests do not
include the authentication headers we [are planning to] use, so an
attacker would need to directly compromise the credentials from the
user's browser, not just fire off a request.

[1] https://en.wikipedia.org/wiki/Cross-origin_resource_sharing
[2] https://expressjs.com/en/resources/middleware/cors.html

Issue #96 Authentication
Issue PhilanthropyDataCommons/data-viewer#6 Proof of concept
      authentication via Keycloak