Rename auth variables
In anticipation of the change from `x-api-key: [secret]` to a standard
`authorization: Bearer [JSON Web Token]` header, rename the variable
in the test code, routers, logger, and documentation to one that
covers both cases. Having this change in a separate commit makes it
easier to see substantive differences in following commits.

Issue #96 Authentication