Validate Bearer JWT for authentication
Expect a JSON Web Token (JWT) in the `Authorization` header for
protected resources. Validate the JWT using the configured realm key.
This allows the standard OpenID Connect protocol to be used for
authentication to an auth server, namely Keycloak, and the resulting
token to be valid for calls to this service. For example, a user of a
browser-based UI can log in using the standard Authorization Code flow
and that UI can make calls to this back-end service using the JWT.

For audit trail purposes, it is still important to log some portion of
the JWT. A JWT, while expected to be short-lived, is still useful as
credentials in its complete form. Part of the validation of a JWT is
signature validation. To keep the audit trail while preventing replay
from a server log, redact the signature of the JWT in logs.

The OpenAPI documentation is a normal browser-based Client. This
change expects the documentation Client to have a particular name,
'pdc-openapi-docs', to be configured in the 'pdc' realm of the
authentication service.

Issue #96 Authentication