Wrap proposal collections in a Bundle object
We want to be able to provide metadata related to queries, and a wrapper
object enables that.  For now the object is starting with the most
simple of metadata (the "count" of entities).

The structure of this bundle is inspired by the FHIR `Bundle`
concept[1], though we are making it our own (for instance `total` is not
associated with the concept of search in our API, wherease for FHIR
total only exists when a search is being performed).

[1] https://www.hl7.org/fhir/bundle-definitions.html

Issue #179 Paginate API results