Add data seed helper script 2
Developers or users of the API can use this script to add base fields
to their local databases. It is intended to run after the first
script. In other words, even if you are just starting out, run the
script starting with `0001` followed by `0002` and so forth.

This script ran on the pilot database around 2023-04-19T15:08-0600.
No errors occurred.

Issue #102 Make canonical fields and application forms from spreadsheets