Validate base field `dataType` values as JSON
The goal is to use the values of `dataType` for a given base field to
validate values of `value` for given proposal field values. The change
in this commit is a stepping stone along the way to that goal by first
validating the value of `dataType`. We would like `dataType` to be a
valid JSON Schema validation value. An example validation value is
`{ "type": "string", "format: "date" }`. See other valid values at
https://json-schema.org/understanding-json-schema/reference/string.html#dates-and-times

In order to validate the `dataType` values, we can first exploit the
`jsonb` postgresql type. The database itself will therefore validate
that the value for `dataType` is JSON. All JSON Schema values are JSON
but not all JSON values are JSON Schemas. At the application level, we
use Ajv to validate that the `dataType` is not only valid JSON but
also a valid JSON Schema.

Issue #308 Validate proposal field values using base field JSON Schema type