WIP Lint SQL with SQLFluff
One of the great advantages to using tinypg and postgres-migrations is
the ability to separate our SQL queries and migrations into SQL files.
We've already been benefiting from the editor support that approach
gains us!

Another benefit is the ability to use static analysis to improve our SQL
coding practices. Start using SQLFluff[1], a SQL linter that works with
PostgreSQL-flavored SQL.

WIP because it still needs:
- dev docs
- CI
- configuration
- to fix existing violations
- to ignore violations in existing migrations, since those can't be
  changed
- team consent

[1] https://docs.sqlfluff.com/en/stable/index.html