Lint SQL with SQLFluff
One of the great advantages to using tinypg and postgres-migrations is
the ability to separate our SQL queries and migrations into SQL files.
We've already been benefiting from the editor support that approach
gains us!

Another benefit is the ability to use static analysis to improve our SQL
coding practices. Start using SQLFluff[1], a SQL linter that works with
PostgreSQL-flavored SQL.

I looked at other options, and SQLFluff is the best in this space. It
does use Python, and introducing another language to our dev
dependencies is unfortunate, but I think it's worthwhile.

Add a basic configuration that ignores all our existing errors, and all
the migrations (since they cannot be changed). This allows SQLFluff to
run, and prevent any errors we don't already have. Rule violations will
be fixed in subsequent commits.

Run SQLFluff as a new lint job in GitHub Actions, so that pull requests
will be checked.

[1] https://docs.sqlfluff.com/en/stable/index.html