Signal node version to npm
Dependabot uses `npm outdated` when checking what dependencies need
updating.  We want npm to be aware of the node version so that it will
not recommend updates that are incompatible with the current version of
node.

I looked for ways to avoid having to specify the engine in two places
(`.node-version` as well as `package.json`) but couldn't find a clear
solution.