Add data platform provider endpoints
These endpoints allow the PDC to store raw data that has been pulled
from data platform providers.  For now this data is completely
disassociated with the rest of the PDC and we are simply acting as
a cache for the data viewer client.

Eventually we may bypass the PDC entirely for this data viewer use case
OR we may store the data in a way that integrates with the PDC.

Issue #368 Create temporary API for storing data platform provider data