Remove id requirement from table metrics
We have a helper function for collecting table metrics; that function
relied on there being a generated ID field in all tables but nothing was
actually using that value.  We want to introduce a table with a natural
primary key instead of an `id` and we'll still need to be able to use
this metrics query.