Add data seed helper scripts 3 and 4
Developers or users of the API can use this script to add base fields
to their local databases. It is intended to run after the preceding
scripts. In other words, even if you are just starting out, run the
script starting with `0001` followed by `0002` and so forth.

Issue #102 Make base fields and application forms from spreadsheets