Defer to @tsconfig settings for module
The "module" setting is defined in the base tsconfig.json that we're
using, and overriding it actually resulted in a mismatch with the
latest `moduleResolution` that was being defined in that base.

The `Node16` setting appears to simply signal *support* for ESM
depending on the `type` setting in `package.json`.  This is based on the
ECMAScript Modules in Node.js documentation[1] as well as the TS
configuration documentation[2].

I've added `"type": "commonjs"` to `package.json` although the omitted
setting defaulted to `commonjs` which means this change is just to make
the setting explicit.

[1] https://www.typescriptlang.org/docs/handbook/esm-node.html
[2] https://www.typescriptlang.org/tsconfig#node16nodenext-nightly-builds