Add bulk upload template
This is the first draft of a bulk upload template which will eventually
be used as part of a bulk upload processing.  The template is expected
to change and evolve over time.

The fields listed represent a subset of the current base fields, and the
most important aspect of the template is the "code" for each column,
which is what the backend will use to map values to base fields.

Issue #561 Create the bulk upload CSV template