Explicitly reference enum values
The new linting rules called out that the `errorCode` value being
processed was a generic string but we were invoking switch cases against
an enum type (`PostgresErrorCode`).

This worked because enums are weird in javascript, but in order to be
more explicit we need to compare with the `valueOf()` the enum.

In discussion about this code the idea came up that maybe we should be
throwing errors with HTTP codes / messages provided in the first place
(so we could simplify this middleware).  That is out of scope for this
commit but would be reasonable future work to explore.