Allow schema specification when migrating
The upstream copy of our postgres-migrations project has an issue where
it does not support postgres schemas, and if ANY schema has a migrations
table then migrations will fail.  This was frustrating for running
tests, but recently while looking into developing a job queue which
would require a separate schema it became a blocking issue.

Upstream also appears to no longer be maintained (the most recent commit
as of this writing was about 2 years ago, and the PR we opened to fix
the schema limitation has been open for around 18 months).

Fortunately there is a fork of the library designed specifically for
adding schema support.

This replaces the library with the modified version and also updates our
migration code to utilize the schema functionality.

Issue #609 Consider finding a new tool for migrations