Update basefield test insertions to use helpers
We have raw SQL baked into our tests, which made sense at the time, but
it ends up being brittle when we want to adjust tables since it requires
manual editing of a ton of SQL.  Luckily we already have a great db
query tool that we can use to DRY out this aspect of the tests.

This change is being made only to base fields because we are about to
edit the table structure of base fields for issue #612.