Add prefix to presigned posts
The presigned post API determines the key of the object that the client
will upload to S3. Previously, the intent was to have separate buckets
for processed and unprocessed files, but we later decided to have only a
single bucket for both. That change was partially implemented by commit
b78df169244196e2d980b1de6cac6f32a9c14217, but that commit did not update
the key generation.

Add a prefix to the object keys generated by the presigned post API,
both to keep the objects in our bucket better organized, and to allow us
to (later) implement a object lifecycle policy that will automatically
delete old unprocessed objects.

Issue #564 Create "buckets" on DigitalOcean for file upload
Issue #597 Expire unregistered uploads