Replace sourceUrl with sourceKey
Now that we're implementing the bulk upload processor it's quickly
apparent that we do NOT want to take an arbitrary URL as the "source" of
a bulk upload.  There are two reasons for this:

1. The bulk upload flow involves an S3 presigned post, which is not
   going to be publicly readable and so we will want to access it via an
   S3 key using credentials and the s3 client.
2. Arbitrary URLs might open us up to various attack vectors (although
   we are still allowing arbitray content to be uploaded to S3 so it
   might be moot)

Regardless, there is no short term use case for accepting arbitrary
URLs, and in order to support the S3 workflow we are going to need to
know the key (not the bucket, since there is only one PDC bucket).

This removes the URL and replaces it with the key.

Issue #559 Set up a queue / processing system for handling bulk proposal uploads