Implement source download for bulk upload
The first thing we need to do in order to process a bulk upload is
download the content of the file to temporary storage so that we can
read it.  We're only supporting the download of files that have been
uploaded directly to our S3 bucket using the presigned post workflow.

It's critical that the temporary file be cleaned up *exactly once*.

Issue #559 Set up a queue / processing system for handling bulk proposal uploads