Create entities when processing bulk upload
When a bulk upload is processed we treat the bulk upload itself as a new
opportunity, application form, and set of application form fields for
the bulk upload itself.  We also create one new proposal (and set of
proposal field values) for each row / record in the csv.

Issue #559 Set up a queue / processing system for handling bulk proposal uploads