Rename BaseFieldCreate to BaseFieldWrite
Naming sure is fun.

Some of our types have `Create` varieties while others have `Write`
varieties. I don't know which was more recent, but now that we have an
update endpoint I think it makes more sense to lean into `Write`.

The rationale for `Write` is that `Create` and `Update` are both write
operations, and they are going to require the same fields. The
alternative would mean two identical types / type guards / schemas (one
for create, one for write).

Another reason to just have the one is that our swagger specification
has a single type with "readonly" attributes, and essentially
"generates" a read vs write version of the schema.